backend/requirements.txt
fastapi==0.128.8
uvicorn[standard]==0.39.0
python-dotenv==1.2.1
pdfplumber==0.11.8
pytesseract==0.3.13
pdf2image==1.17.0
Pillow==11.3.0
numpy==2.0.2
faiss-cpu==1.13.0
langchain==0.3.27
langchain-openai==0.3.35
langchain-community==0.3.31
openai==2.24.0
pydantic-settings==2.11.0
python-multipart